tray functionality moved to separate file, app first instance is activated when the second instance is tried to be run
