+++ .gitignore
@@ -7,6 +7,7 @@
 /app-builds
 /release
 main.js
+tray.js
 src/**/*.js
 !src/karma.conf.js
 *.js.map
